pyproject.toml
[tool.poetry]
name = "excelrag"
version = "0.1.0"
description = "Chat with your data, compose emails and create tasks from the insights"
authors = ["bcx <femolak@gmail.com>"]
license = "GPLv2"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.7"
langchain-experimental = "^0.3.3"
langchain-openai = "^0.2.5"
streamlit = "^1.39.0"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
